requirements.txt
# Core RAG deps (provided in rag/requirements.txt)
# We re-export here for a single pip install step
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0
numpy>=1.26.0

# Voice
SpeechRecognition>=3.10.0
pyaudio>=0.2.14; platform_system != "Darwin"  # macOS often uses alternative input devices
PyAudio>=0.2.14; platform_system == "Darwin"
gTTS>=2.5.1

# CLI
argparse>=1.4.0
fastapi>=0.111.0
uvicorn>=0.30.0
rank-bm25>=0.2.2
openai>=1.30.0
requests>=2.31.0
llama-cpp-python>=0.2.86
langchain>=0.2.7
langchain-community>=0.2.6
langchain-openai>=0.1.20
pydantic>=2.7.0